[优化] UiStateTestUnique增大参数范围

+++ annotation/src/main/java/com/mock/annotation/unique/UiStateTestUnique.kt
@@ -9,7 +9,7 @@ package com.mock.annotation.unique
 @Retention(value = AnnotationRetention.SOURCE)
 @Target(AnnotationTarget.FIELD)
 annotation class UiStateTestUnique(
-    val intDef: IntArray = [],
+    val intDef: LongArray = [],
     val stringDef: Array<String> = [],
     val classDef: Array<String> = [],
 )
